Updates based on code review.
