+++ projects/packages/search/src/instant-search/components/answers-panel.scss
@@ -1,4 +1,5 @@
 @use "../lib/styles/helper";
+@use "@wordpress/base-styles/mixins" as wp-mixins;
 
 .jp-search-answers-panel {
 	background: #fff;
@@ -31,7 +32,7 @@
 		margin-right: helper.$results-margin-md;
 	}
 
-	@include helper.break( "<m" ) {
+	@include wp-mixins.break-small() {
 		margin-left: helper.$results-margin-sm;
 		margin-right: helper.$results-margin-sm;
 	}
